+++ backend/canisters/local_user_index/impl/src/jobs/mod.rs
@@ -8,6 +8,7 @@ pub mod delete_users;
 pub mod moderate_messages;
 pub mod pull_daily_puzzle;
 pub mod refund_cycles;
+pub mod start_user_migrations;
 pub mod topup_canister_pool;
 pub mod topup_canisters;
 pub mod upgrade_communities;
@@ -22,6 +23,7 @@ pub(crate) fn start(state: &RuntimeState) {
     moderate_messages::start_job_if_required(state);
     pull_daily_puzzle::start_job();
     refund_cycles::start_job_if_required(state, None);
+    start_user_migrations::start_job_if_required(state);
     topup_canister_pool::start_job_if_required(state, None);
     topup_canisters::start_job();
     upgrade_communities::start_job_if_required(state);