+++ backend/canisters/user_index/impl/src/model/mod.rs
@@ -16,6 +16,7 @@ pub mod storage_index_users_to_remove_batch;
 pub mod streak_insurance_logs;
 pub mod user;
 pub mod user_map;
+pub mod user_migrations;
 pub mod users_last_online;
 
 pub const MAX_AVATAR_SIZE: usize = 250_000;